Added pagination

+++ backend/src/main/java/com/arpan/backend/repository/UserRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<Users, Integer> {
+public interface UserRepo extends JpaRepository<Users, Long> {
 
     Optional<Users> findByEmail(String email);
